+++ test/test_helper.rb
@@ -1,9 +1,9 @@
 # Configure Rails Environment
 ENV["RAILS_ENV"] = "test"
 
-require 'minitest/autorun'
 require File.expand_path("../dummy/config/environment.rb", __FILE__)
-require "rails/test_help"
+require 'rails/test_help'
+require 'minitest/rails'
 
 ServiceBase.load_services